+++ tests/mutation/test_mutation.py
@@ -28,7 +28,7 @@ from mutmut.__main__ import git_head
 from mutmut.__main__ import git_tracked_non_py_files
 from mutmut.__main__ import run_forced_fail_test
 from mutmut.configuration import Config
-from mutmut.configuration import HotForkWarmup
+from mutmut.configuration import ForkServerWarmup
 from mutmut.configuration import ProcessIsolation
 from mutmut.core import MutmutProgrammaticFailException
 from mutmut.mutation.data import MutantLineSpans
@@ -1627,8 +1627,8 @@ def _config_for_invalidation(**overrides):
         on_dependency_change="warn",
         use_git_change_detection=True,
         process_isolation=ProcessIsolation.FORK,
-        hot_fork_warmup=HotForkWarmup.COLLECT,
-        max_orchestrator_restarts=3,
+        forkserver_warmup=ForkServerWarmup.COLLECT,
+        max_forkserver_restarts=3,
         preload_modules_file=None,
         log_to_file=False,
         log_file_path="mutants/mutmut-debug.log",